+++ clouddriver-kubernetes/src/main/java/com/netflix/spinnaker/config/KubernetesConfiguration.java
@@ -17,6 +17,8 @@ package com.netflix.spinnaker.config;
 
 import com.netflix.spectator.api.Registry;
 import com.netflix.spinnaker.clouddriver.kubernetes.caching.KubernetesProvider;
+import com.netflix.spinnaker.clouddriver.kubernetes.config.KubernetesAccountProperties;
+import com.netflix.spinnaker.clouddriver.kubernetes.config.KubernetesAccountProperties.ManagedAccount;
 import com.netflix.spinnaker.clouddriver.kubernetes.config.KubernetesConfigurationProperties;
 import com.netflix.spinnaker.clouddriver.kubernetes.health.KubernetesHealthIndicator;
 import com.netflix.spinnaker.clouddriver.kubernetes.security.KubernetesCredentials;
@@ -38,6 +40,7 @@ import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
@@ -51,6 +54,19 @@ public class KubernetesConfiguration {
   @ConfigurationProperties("kubernetes")
   public KubernetesConfigurationProperties kubernetesConfigurationProperties() {
     return new KubernetesConfigurationProperties();
+  }
+
+  @Bean
+  @RefreshScope
+  @ConfigurationProperties("kubernetes")
+  @ConditionalOnProperty(
+      prefix = "kubernetes",
+      name = "custom-property-binding-enabled",
+      havingValue = "false",
+      matchIfMissing = true)
+  @Primary
+  public KubernetesAccountProperties kubernetesAccountProperties() {
+    return new KubernetesAccountProperties();
   }
 
   @Bean
@@ -70,15 +86,13 @@ public class KubernetesConfiguration {
       value = KubernetesNamedAccountCredentials.class,
       parameterizedContainer = AbstractCredentialsLoader.class)
   public AbstractCredentialsLoader<KubernetesNamedAccountCredentials> kubernetesCredentialsLoader(
-      @Nullable
-          CredentialsDefinitionSource<KubernetesConfigurationProperties.ManagedAccount>
-              kubernetesCredentialSource,
-      KubernetesConfigurationProperties configurationProperties,
+      @Nullable CredentialsDefinitionSource<ManagedAccount> kubernetesCredentialSource,
+      KubernetesAccountProperties accountProperties,
       KubernetesCredentials.Factory credentialFactory,
       CredentialsRepository<KubernetesNamedAccountCredentials> kubernetesCredentialsRepository) {
 
     if (kubernetesCredentialSource == null) {
-      kubernetesCredentialSource = configurationProperties::getAccounts;
+      kubernetesCredentialSource = accountProperties::getAccounts;
     }
     return new BasicCredentialsLoader<>(
         kubernetesCredentialSource,
@@ -97,7 +111,7 @@ public class KubernetesConfiguration {
 
   @Bean
   @ConditionalOnMissingBean(
-      value = KubernetesConfigurationProperties.ManagedAccount.class,
+      value = ManagedAccount.class,
       parameterizedContainer = CredentialsDefinitionSource.class)
   public CredentialsInitializerSynchronizable kubernetesCredentialsInitializerSynchronizable(
       AbstractCredentialsLoader<KubernetesNamedAccountCredentials> loader) {
